+++ src/GNR2_Firmware/revA/grn2_ra6t2/FTEX/src/peripherals_abstraction/pulse_frequency_gpt.c
@@ -103,7 +103,7 @@ uint64_t PulseFrequency_InputCaptureValue_GPT (WheelFrequency_Handle_GPT_t * pHa
 /**
   Function used to update the capture variables from the Genral Purpose Timer interrupt
 */
-void PulseFrequency_IsrCallUpdate_GPT(WheelFrequency_Handle_GPT_t * pHandle , uint64_t wCapture)
+void PulseFrequency_IsrCallUpdate_GPT(WheelFrequency_Handle_GPT_t * pHandle , uint32_t wCapture)
 {
 	/* Capture the count in a variable */
 	pHandle->gpt_capture_count = wCapture;
